+++ features/admin.users.v1/utils/user-management-utils.ts
@@ -81,39 +81,6 @@ export class UserManagementUtils {
     }
 
     /**
-     * Generates code verifier for PKCE.
-     *
-     * @returns A generated code verifier.
-     */
-    public static generateCodeVerifier(): string {
-        const array: Uint8Array = new Uint8Array(32);
-
-        window.crypto.getRandomValues(array);
-
-        return btoa(String.fromCharCode(...array))
-            .replace(/\+/g, "-")
-            .replace(/\//g, "_")
-            .replace(/=+$/, "");
-    };
-
-    /**
-     * Generates code challange for code verifier.
-     *
-     * @param codeVerifier - The code verifier.
-     * @returns A generated code challange.
-     */
-    public static getCodeChallangeForTheVerifier = async (codeVerifier: string) => {
-        const encoder: TextEncoder = new TextEncoder();
-        const encodedCodeVerifier: Uint8Array = encoder.encode(codeVerifier);
-        const hashedCodeVerifier: ArrayBuffer = await window.crypto.subtle.digest("SHA-256", encodedCodeVerifier);
-
-        return btoa(String.fromCharCode(...new Uint8Array(hashedCodeVerifier)))
-            .replace(/\+/g, "-")
-            .replace(/\//g, "_")
-            .replace(/=+$/, "");
-    };
-
-    /**
      * Resolves the sub header of the user list item.
      *
      * @param user - User object.
