+++ modules/i18n/src/models/namespaces/user-ns.ts
@@ -89,7 +89,6 @@ export interface userNS {
             };
         };
         userActionZoneGroup: {
-            header: string;
             impersonateUserZone: {
                 actionTitle: string;
                 header: string;
