+++ modules/react-components/src/components/danger-zone/danger-zone-group.tsx
@@ -55,7 +55,7 @@ export const DangerZoneGroup: FunctionComponent<PropsWithChildren<DangerZoneGrou
     const defaultClassName = className ?? "danger-zone";
 
     return (
-        <div className= { (className ?? defaultClassName) + "-wrapper" }>
+        <div className= { defaultClassName + "-wrapper" }>
             <Header
                 as="h5"
                 className="bold-text"
